Métodos separados em suas respectivas classes. Autor: todos.
