+++ src/main/java/controller/PerfilController.java
@@ -1,0 +1,161 @@
+package controller;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import model.acesso.Perfil;
+import model.acesso.PerfilDAO;
+import model.acesso.Permissao;
+
+public class PerfilController {
+	
+	private static PerfilController instance;
+
+	private PerfilController() {
+	}
+
+	public static PerfilController getInstance() {
+		if (instance == null) {
+			instance = new PerfilController();
+		}
+		return instance;
+	}
+	
+	/** PerfilController
+	 * Vai ser feita uma consulta no banco de dados pelo ID do perfil para saber se o {@link Perfil} 
+	 * tem permissoes.
+	 * 
+	 * @param perfil - Perfil
+	 * @param permissao - Permissao
+	 * @return ArrayList<Permissao>
+	 */
+	public boolean verificarPermissao(Perfil perfil, Permissao permissao) {
+		List<Permissao> listaDePermissoesDessePerfil = listarPermissoesDeUmPerfil(perfil.getIdPerfil());
+		if (listaDePermissoesDessePerfil.contains(permissao)) {
+			return true;
+		} else {
+			return false;
+		}
+	}
+
+	/** PerfilController
+	 * Criacao de um {@link Perfil} no objeto.
+	 * 
+	 * Responsavel por criar um objeto do tipo {@link Perfil} com o atributo
+	 * nomePerfil. O objeto {@link Perfil} e enviado ao {@link PerfilDAO} para ser
+	 * persistido no banco de dados.
+	 * 
+	 * @param String - nomePerfil
+	 * 
+	 */
+	public void criarPerfil(String nomePerfil) {
+		Perfil perfil = new Perfil(nomePerfil);
+		PerfilDAO.getInstance().criar(perfil);
+	}
+
+	/** PerfilController
+	 * Alteracao de um {@link Perfil}.
+	 * 
+	 * Responsavel por alterar um {@link Perfil} pre existente com os atributos
+	 * idPerfil e nomePerfil. A busca e realizada pelo idPerfil por intermedio do
+	 * PerfilDAO,o nome do perfil e adicionado ao objeto e alterado no banco de
+	 * dados.
+	 * 
+	 * @param Integer idPerfil
+	 * @param String  nomePerfil
+	 */
+	public void alterarPerfil(Integer idPerfil, String nomePerfil) {
+		Perfil perfil = PerfilDAO.getInstance().consultarPorId(idPerfil);
+		perfil.setNomePerfil(nomePerfil);
+		PerfilDAO.getInstance().alterar(perfil);
+	}
+
+	/** PerfilController
+	 * RemoÃƒÂ§ÃƒÂ£o de um {@link Perfil} pelo id.
+   *
+	 * ResponsÃƒÂ¡vel por deletar um objeto do tipo {@link Perfil} com os atributos idPerfil.
+	 * O objeto {@link Perfil} ÃƒÂ© enviado ao {@link Perfil} para ser removido no banco de dados.
+	 * 
+	 * @param idPerfil - Integer
+	 */
+	public void deletarPerfil(Integer idPerfil) {
+		Perfil perfil = PerfilDAO.getInstance().consultarPorId(idPerfil);
+		PerfilDAO.getInstance().deletar(perfil);
+	}
+
+	/** PerfilController
+	 * Consulta de {@link Perfil} pelo id.
+	 * 
+	 * Responsavel por alterar um {@link Perfil} pre existente com os atributos
+	 * idPerfil e nomePerfil. A busca e realizada pelo idPerfil por intermedio do
+	 * {@link PerfilDAO},o nome do perfil e adicionado ao objeto e alterado no banco de
+	 * dados.
+	 * 
+	 * 
+	 * @param Integer 
+	 * @throws NullPointerException Caso nÃƒÂ£o exista o {@link Perfil} no banco de dados.
+	 * @return Perfil
+	 */
+	public Perfil consultarPerfil(Integer idPerfil) {
+		try {
+			return PerfilDAO.getInstance().consultarPorId(idPerfil);
+		} catch (NullPointerException e) {
+			return null;
+		}
+	}
+
+	/** PerfilController
+	 * Tem a funcao de consultar por nome um objeto do tipo {@link Perfil} com o atributo nomePerfil.
+	 * O objeto {@link Perfil} vai ser consultado no banco de dados pelo nome.
+	 * 
+	 * @param nome - String
+	 * @return {@link Perfil} 
+	 */
+	public Perfil consultarPerfil(String nome) {
+		try {
+			return PerfilDAO.getInstance().consultarPorNome(nome);
+		} catch (NullPointerException e) {
+			return null;
+		}
+
+	}
+
+	/** PerfilController
+	 * Consulta todos os perfis no banco de dados.
+	 * 
+	 * @return resultado - ArrayList<Perfil>
+	 */
+	public ArrayList<Perfil> listarTodosOsPerfis() {
+		ArrayList<Perfil> perfisEncontrados = (ArrayList<Perfil>) PerfilDAO.getInstance().listar();
+		ArrayList<Perfil> resultado = !perfisEncontrados.isEmpty() ? perfisEncontrados : null;
+		return resultado;
+
+	}
+
+	/** PerfilController
+	 * Lista todas as {@link Permissao} de um {@link Perfil}.
+	 * 
+	 * O método irá pesquisar todas as {@link Permissao} que o {@link Perfil} possui fazendo uma busca
+	 * pelo idPerfil, o qual será enviado ao {@link PerfilDAO}. O retorno será todas as {@link Permissao}
+	 * que um {@link Perfil} possui.
+	 * 
+	 * @param idPerfil
+	 * @return List<Permissao>
+	 */
+	public List<Permissao> listarPermissoesDeUmPerfil(int idPerfil) {
+		return PerfilDAO.getInstance().listarPermissoes(idPerfil);
+	}
+
+	/** PerfilController
+	 *  Ã‰ atribuido uma {@link Permissao} a um {@link Perfil} quando Ã© passado uma {@link Permissao} 
+	 *  para aquele {@link Perfil}.
+	 * 
+	 * @param permissao Permissao
+	 * @param perfil Perfil
+	 */
+	public void atribuirPermissaoAUmPerfil(Permissao permissao, Perfil perfil) {
+		PerfilDAO.getInstance().atribuirPermissaoAUmPerfil(perfil, permissao);
+	}
+
+
+}
