+++ src/test/java/controller/PerfilControllerTest.java
@@ -1,0 +1,14 @@
+package controller;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+public class PerfilControllerTest {
+
+	@Test
+	public void test() {
+		fail("Not yet implemented");
+	}
+
+}
