+++ src/test/java/controller/PermissaoControllerTest.java
@@ -1,0 +1,14 @@
+package controller;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+public class PermissaoControllerTest {
+
+	@Test
+	public void test() {
+		fail("Not yet implemented");
+	}
+
+}
