+++ src/test/java/controller/RotinasControllerTest.java
@@ -1,0 +1,14 @@
+package controller;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+public class RotinasControllerTest {
+
+	@Test
+	public void test() {
+		fail("Not yet implemented");
+	}
+
+}
